Frontend Completed
Created UseStates and UseEffects for acquiring, storing and displaying information on appropriate category pages, obtained from the post and get methods from the backend.

Most SQL query data obtained is stored in Lists which are then mapped and displayed to the relevant page based on the information we want to portray. For example, BreadPQList contains a list of all the data of the bread item which is available in only Downtown Malls.

Appropriate changes made to Bakery.jsx, Beverage.jsx, Cart.jsx, Dairy.jsx, Fruits.jsx, Meat.jsx and Vegetables.jsx.